// src/types.ts
export type FiltersProps = {
  isAlcoholic: boolean;
  isNonAlcoholic: boolean;
};

export type DataProps = {
  idDrink: string;
  strAlcoholic: string;
  strDrink: string;
  strInstructions: string;
  strIngredients: string[];
};

export type OriginalDataProps = {
  [key: string]: string;
  idDrink: string;
  strAlcoholic: string;
  strDrink: string;
  strInstructions: string;
  strIngredient1: string;
  strIngredient2: string;
  strIngredient3: string;
  strIngredient4: string;
  strIngredient5: string;
  strIngredient6: string;
  strIngredient7: string;
  strIngredient8: string;
  strIngredient9: string;
  strIngredient10: string;
  strIngredient11: string;
  strIngredient12: string;
  strIngredient13: string;
  strIngredient14: string;
  strIngredient15: string;
};